requirements.txt
qq-botpy>=1.2.0
anthropic>=0.40.0
python-dotenv>=1.0.0
textual>=0.80.0
